#1404: remove internal interfaces, fix build errors
