+++ cli/src/main/java/com/devonfw/tools/ide/io/ini/IniProperty.java
@@ -1,17 +1,44 @@
 package com.devonfw.tools.ide.io.ini;
 
 /**
- * Container for a property in an {@link IniFile}
+ * Implementation of {@link IniProperty}
  */
-abstract class IniProperty extends IniElement {
+public class IniProperty extends IniElement {
+
+  private final String key;
+  private final String value;
+
+  /**
+   * creates a new IniPropertyImpl with indentation level 0
+   *
+   * @param content String representation of the property in the file
+   */
+  public IniProperty(String content) {
+    this.setContent(content);
+    int index = content.indexOf('=');
+    if (index <= 0) {
+      throw new IllegalArgumentException("Ini Property declaration must contain a '=' symbol");
+    }
+    key = content.substring(0, index).trim();
+    value = content.substring(index + 1).trim();
+  }
 
   /**
    * @return property key
    */
-  public abstract String getKey();
+  public String getKey() {
+    return key;
+  }
 
   /**
    * @return property value
    */
-  public abstract String getValue();
+  public String getValue() {
+    return value;
+  }
+
+  @Override
+  public String toString() {
+    return this.getContent();
+  }
 }
